html/browsers/browsing-the-web/back-forward-cache/service-worker-unregister.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

promise_test(async t => {
  // Register a service worker and make this page controlled.
  const workerUrl =
      'resources/service-worker.js?pipe=header(Service-Worker-Allowed,../)';
  const registration =
      await service_worker_unregister_and_register(t, workerUrl, './');
  t.add_cleanup(_ => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');
  const controllerChanged = new Promise(
      resolve => navigator.serviceWorker.oncontrollerchange = resolve);
  await claim(t, registration.active);
  await controllerChanged;

  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());

  const urlA = location.origin + executorPath + pageA.context_id;
  const urlB = originCrossSite + executorPath + pageB.context_id;

  // Open `urlA`.
  window.open(urlA, '_blank', 'noopener');
  await pageA.execute_script(waitForPageShow);

  // Navigate to `urlB`.
  await pageA.execute_script(
    (url) => prepareNavigation(() => {
      location.href = url;
    }),
    [urlB]);
  await pageB.execute_script(waitForPageShow);

  // Unregister the service worker when the controlled `pageA` is in BFCache.
  // This should not crash.
  await registration.unregister();

  // Back navigate and check whether the page is restored from BFCache.
  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );
  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  // `pageA` should be still controlled after restored from BFCache.
  assert_equals(
    await pageA.execute_script(
        () => fetch('/is-controlled').then(r => r.text())),
    'controlled',
    'pageA should be still controlled');

}, 'Unregister service worker while a controlled page is in BFCache');
</script>
